Switched from List<Pair<K,V>> to Map<K,V>
 Changes to be committed:
	deleted:    fabric-commands-v0/src/main/java/net/fabricmc/fabric/api/commands/FabricCommandRegistry.java
	new file:   fabric-commands-v0/src/main/java/net/fabricmc/fabric/api/registry/FabricCommandRegistry.java
	new file:   fabric-commands-v0/src/main/java/net/fabricmc/fabric/impl/command/FabricCommandRegistryImpl.java
	deleted:    fabric-commands-v0/src/main/java/net/fabricmc/fabric/impl/commands/FabricCommandRegistryImpl.java
