+++ fabric-commands-v0/src/main/java/net/fabricmc/fabric/impl/command/FabricCommandRegistryImpl.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.fabricmc.fabric.impl.command;
+
+import net.minecraft.command.AbstractCommand;
+
+import net.fabricmc.fabric.api.registry.FabricCommandRegistry;
+
+public class FabricCommandRegistryImpl implements FabricCommandRegistry {
+	@Override
+	public void register(AbstractCommand command, boolean dedicated) {
+		FABRIC_COMMANDS.put(command, dedicated);
+	}
+}
